GraphBuilder: optimization, pass IdIndexMap to built graph instead of clone
This optimization is usefull when a builder is used to construct a single
graph and then discarded. Previously, the IdIndexMap was cloned to the built
graph. After the change, if an immutable graph (without weights due to impl
reasons) is constructed, the maps (one of vertices and one of edges) are
passed to the graph without copying and the builder remaing without the maps.
A reference to the graph is saved, such that if the builder is used again the
maps are first copied from the graph back to the builder, and then the
operations continue as normal.