IndexGraphBuilder: optimization, pass endpoints array to built graph instead of clone
This optimization is usefully in cases the IndexGraphBuilder is used only
once to construct a single graph and than it is discarded. In such cases,
the builder allocate it ownss memory for the edges endpoints, and than
a new array is allocated for the endpoints stored in the newly created
graph.